requirements.txt
aiohttp==2.2.5
alembic==0.9.5
pyquery==1.2.17
python-telegram-bot==8.0
SQLAlchemy==1.1.14
python-dotenv==0.7.1
emoji==0.4.5